requirements.txt
# Core Web Framework (HuggingFace Spaces optimized)
streamlit>=1.46.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Essential Scientific Computing
numpy>=1.23.0
pandas>=1.4.0
scipy>=1.9.0

# Machine Learning & NLP (Memory optimized)
torch>=2.0.0,<2.5.0  # TODO: upgrade to >=2.6.0 (5 CVEs) after full test pass
sentence-transformers>=2.2.0,<3.0.0
transformers>=4.30.0,<5.0.0
tokenizers>=0.13.0

# ML Operations & Experiment Tracking
mlflow>=2.9.0,<3.0.0  # Experiment tracking. TODO: upgrade to >=3.1.0 (4 CVEs) — major version, API changes expected

# Vector Search & Indexing
faiss-cpu>=1.7.4

# Document Processing (Essential only)
PyMuPDF>=1.23.0

# Text Processing (Core only)
nltk>=3.8.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.4.0
pydantic>=2.0.0,<3.0.0  # Data validation and settings management
pyyaml>=6.0.0  # YAML configuration parsing

# Epic 2 Graph Processing (Lightweight)
networkx>=3.0,<4.0.0
spacy>=3.7.0,<3.8.0

# HuggingFace API Integration (Essential for cloud deployment)
huggingface-hub>=0.16.0

# Visualization (Lightweight)
plotly>=5.17.0,<6.0.0

# Epic 1 Multi-Model Integration
openai>=1.0.0  # Official OpenAI Python client
mistralai>=0.4.0  # Official Mistral Python client  
tenacity>=8.0.0  # Retry logic with exponential backoff
tiktoken>=0.5.0  # OpenAI token counting
mistral-common>=1.0.0  # Mistral token counting

# Epic 8 Cloud-Native Dependencies
redis>=5.0.1  # Redis Python client with async support for cache service

# Testing & Quality Assurance
pytest>=7.4.0,<8.0.0  # Testing framework
pytest-cov>=4.1.0  # Coverage reporting
pytest-xdist>=3.3.0  # Parallel test execution
pytest-asyncio>=0.21.0  # Async test support

# RAG Evaluation
ragas>=0.1.0  # RAG evaluation framework with quality metrics

# Optional dependencies for local development only
# accelerate>=0.20.0  # Uncomment for local GPU acceleration
# pdfplumber>=0.10.0  # Uncomment for advanced PDF parsing
# fastapi>=0.104.0    # Uncomment for API server
# uvicorn>=0.24.0     # Uncomment for API server
# ollama>=0.5.0       # Uncomment for local LLM
# python-igraph>=0.10 # Uncomment for advanced graph features

# Language model (will be downloaded automatically)
# en-core-web-sm>=3.7  # Installed via spacy download
# Epic 5 Agent System Dependencies
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5  # For OpenAI agent examples
